IDEA-322117 Convert InspectionProfilerDataHolder to non-light service
This converts InspectionProfilerDataHolder from light service to
non-light service. This allows for a different implementation of the service in Android Studio.
